+++ src/layers/mod.rs
@@ -54,6 +54,8 @@ pub struct LayerData {
     pub tint_color: Option<Color>,
     /// The layer's custom properties, as arbitrarily set by the user.
     pub properties: Properties,
+    /// The layer's type, which is arbitrarily setby the user.
+    pub user_type: Option<String>,
     layer_type: LayerDataType,
 }
 
@@ -76,7 +78,19 @@ impl LayerData {
         reader: &mut impl ResourceReader,
         cache: &mut impl ResourceCache,
     ) -> Result<Self> {
-        let (opacity, tint_color, visible, offset_x, offset_y, parallax_x, parallax_y, name, id) = get_attrs!(
+        let (
+            opacity,
+            tint_color,
+            visible,
+            offset_x,
+            offset_y,
+            parallax_x,
+            parallax_y,
+            name,
+            id,
+            user_type,
+            user_class,
+        ) = get_attrs!(
             for v in attrs {
                 Some("opacity") => opacity ?= v.parse(),
                 Some("tintcolor") => tint_color ?= v.parse(),
@@ -87,8 +101,10 @@ impl LayerData {
                 Some("parallaxy") => parallax_y ?= v.parse(),
                 Some("name") => name = v,
                 Some("id") => id ?= v.parse(),
+                Some("type") => user_type ?= v.parse(),
+                Some("class") => user_class ?= v.parse(),
             }
-            (opacity, tint_color, visible, offset_x, offset_y, parallax_x, parallax_y, name, id)
+            (opacity, tint_color, visible, offset_x, offset_y, parallax_x, parallax_y, name, id, user_type, user_class)
         );
 
         let (ty, properties) = match tag {
@@ -136,6 +152,7 @@ impl LayerData {
             tint_color,
             name: name.unwrap_or_default(),
             id: id.unwrap_or(0),
+            user_type: user_type.or(user_class),
             properties,
             layer_type: ty,
         })